+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to the "azdevops-vscode-simplify" extension will be document
 - Ask for branch to base on when creating a new branch #28
 - Add the possibility to show the parents of the Work Items as well using the setting `azdevops-vscode-simplify.showParentsOfWorkItems`. #11
 - Add the possibility to select multiple Work Items when adding them to the commit message and using the command for that.
+- Add setting `azdevops-vscode-simplify.organizationAndProjectFilter` to filter organizations and projects. #10
 
 ## [0.0.6]